requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
sentence-transformers
faiss-cpu
rank-bm25
numpy
